Configure some leeway in the amount of seconds the IAT can be in the future
It seems like sometimes the JWT we get back from Auth0 has an IAT claim that is 1 second in the future. The Firebase JWT validation logic fails on this, because it expects the IAT to be in the past. However it has a leeway option we can configure, so we can allow the IAT to be x amount of seconds in the future. Set it to 30 for now.